04_drf_test/requirements.txt
# Django + API
django>=4.2
djangorestframework>=3.14

# 图像上传支持
pillow
opencv-python

yomitoku
onnxruntime==1.17.1
numpy==1.26.4

# 文件管理可选（图片存储相关）
python-magic  # 识别上传文件类型
